requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
dateparser==1.0.0
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
pandas==1.3.2
pybind11==2.7.1
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
./whl/tflite_runtime-2.3.1-cp37-cp37m-linux_armv6l.whl; platform_machine == 'armv6l'
--extra-index-url=https://google-coral.github.io/py-repo/
tflite_runtime==2.5.0; platform_machine == 'AMD64'